rename test/e2e_node/build/ to builder/
