+++ test/e2e_node/builder/build.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package build
+package builder
 
 import (
 	"flag"
