+++ test/e2e_node/runner/local/run_local.go
@@ -23,7 +23,7 @@ import (
 	"path/filepath"
 	"strings"
 
-	"k8s.io/kubernetes/test/e2e_node/build"
+	"k8s.io/kubernetes/test/e2e_node/builder"
 
 	"github.com/golang/glog"
 )
@@ -37,13 +37,13 @@ func main() {
 
 	// Build dependencies - ginkgo, kubelet and apiserver.
 	if *buildDependencies {
-		if err := build.BuildGo(); err != nil {
+		if err := builder.BuildGo(); err != nil {
 			glog.Fatalf("Failed to build the dependencies: %v", err)
 		}
 	}
 
 	// Run node e2e test
-	outputDir, err := build.GetK8sBuildOutputDir()
+	outputDir, err := builder.GetK8sBuildOutputDir()
 	if err != nil {
 		glog.Fatalf("Failed to get build output directory: %v", err)
 	}
